backend/requirements.txt
pydantic==2.10.3
langchain==1.1.2
langchain-core==1.1.1
langchain_openai==1.1.0
langchain_community==0.4.1
fastapi==0.118.0
uvicorn==0.37.0
pymongo==4.15.1
dnspython==2.8.0
pymupdf==1.26.5
qdrant-client==1.15.1
langgraph==1.0.2
langgraph-prebuilt==1.0.5
pypdf==6.4.0
fpdf==1.7.2
python-dotenv==1.0.1
python-multipart==0.0.9
typing-extensions==4.12.2
